backend/requirements.txt
fastapi>=0.100.0,<1.0.0
uvicorn[standard]>=0.23.0
pydantic>=2.0.0,<3.0.0
langgraph>=0.0.20
numpy>=1.24.0
sqlalchemy>=2.0.0
httpx>=0.24.0
pytest>=7.4.0
anthropic>=0.25.0
python-dotenv>=1.0.0
openai>=1.0.0
anet-sdk>=1.1.0
